systemd-journal: Prefer SYSLOG_IDENTIFIER over _COMM
In order to not break assumptions, prefer SYSLOG_IDENTIFIER over _COMM.
For example, postfix uses postfix/qmgr as SYSLOG_IDENTIFIER, but _COMM
is only "qmgr". The journal itself uses SYSLOG_IDENTIFIER when
reconstructing the syslog message, so we should not deviate from that
behaviour, either.

Similarly, rsyslog also prefers SYSLOG_IDENTIFIER, so for the sake of
compatibility, doing the same is preferable.

Reported-by: Alexander Görtz
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>