configure.ac: move -Wall and -Wshadow to Makefile.am
Because we have originally set CFLAGS directly, all configure mechanisms
were affected, not just our own codes.
This makes it impossible to enable some of the analysis with -Werror
because autoconf stresses some of the compiler limits on purpose.

Also fixes the omission about -Wshadow not propagating in debug mode.

Signed-off-by: bkil-syslogng <tamas.nagy@balabit.com>